requirements.txt
wheel==0.23.0
dnspython==1.12.0
